Fix package names of FXML resources and open code
Fix two errors in package names:

1. java.lang.IllegalStateException: Location is not set.

The application loads the FXML resources relative to its own package
org.openjfx.hellofx, but the resources are not found because they are in
a different package, org.openjfx. The fix is either to get the resources
using their absolute names, such as "/org/openjfx/primary.fxml", or to
change their package name to org.openjfx.hellofx. This fix changes the
package name.

2. java.lang.module.InvalidModuleDescriptorException:
     Package org.openjfx not found in module

The FXML classes use reflection to load the controllers in
org.openjfx.hellofx, but only the package org.openjfx, which contains no
code, is open to it. Fix the module descriptor to open the correct
package name of org.openjfx.hellofx.

Fixes #39