fails with user testing
Needed a `cd`.
There is no `MainApp`, only an `App`.
More explanation here would be helpful. And please user-test.
Should explain why the `--module-path` is needed when using maven. Maven doesn't put openJfx all in one dir somewhere but makes a forest of versioned dirs, so where would this PATH_TO_FX be? Are we supposed to download openjfx in addition to relying on maven to download it? Confusing.